Refactored JUnit fixture method support to do less reflection and use fewer data structures.
We now do a preliminary pass through all methods and filter for suitability as any kind of fixture method (i.e. zero-arg and void return) instead of checking this for every type of fixture method.

We also no longer build up a structure of suitable methods and then create interceptors based on this structure, avoiding the creation of a map of lists for every spec in the hierarchy regardless of whether it had any fixture methods or not.