SpockConfig can include/exclude by regex pattern string.
The regex pattern is matched against the simple name
of the annotation, spec, or feature.  This addresses
http://code.google.com/p/spock/issues/detail?id=184

Also added to new_and_noteworthy, and updated javadoc in 
IncludeExcludeExtension and RunnerConfiguration.  Removed an apparently 
obsolete part of the example from the latter, because it didn't work 
when I tried it, and there are no specs for it.