Data table cells can reference previous ones
For parametarized tests it could be important for a parameter to depend
on a previously calculated value.

Until now this wasn't possible, because the whole where block was calculated,
before the test method was called (via its generated parameters).
Also, the data providers calculate the whole column,
before moving on to the next one, making the references seem unobvious.

The fix was to provide the data provider methods with the list of already
calculated columns.
And because a cell should only refer to cells in the same row,
the variable references were changed to array indexes (see the getAt part).

This commit should play well with partial parameter definitions.

Partially fixes #509