Fix PollingConditionsSpec
The AppVeyor build is so slow that it causes the spec to fail, since the
initial setup of the polling conditions eventually block takes to long.
By decreasing the initialDelay, this should no longer be an issue.
Furthermore, the timeouts for another test are reduced from seconds to
milliseconds, as there is no reason to waste 5 seconds waiting.