Add ScanScopedBeans-Annotation to support scoped beans better
Currently Spring users run into issues if they have scoped beans, but
the scope is not active during the test start, e.g. RequestScope. This
is now changed, and scoped beans are not scanned by default. To allow
scoped mocks to work, the ScanScopedBeans annotation can be used to
enable scanning for all, or only selected scopes.