align the order of JUnit rule execution with the order of field init
Currently, the order of JUnit rule execution (be it @Rule or @ClassRule) is the reverse order of how the class fields are initialized. This change reverses the order of how the rules are executed to have the same order as class field initialization (assuming that this for every JVM is the same order as returned by Class.getDeclaredFields()).