Fix OOM when comparing large, very different, Strings
Casting calculation of "edit distance memory" in tryReduceStringSizes to
use longs.

The result of multiplying the length of two Strings can be larger than
the maximum value a Integer can hold before overflowing. In this case,
comparing large strings with big differences could cause an overflow.
When it did overflow, the comparison against MAX_EDIT_DISTANCE_MEMORY
would give an incorrect result, possibly causing large Strings to be
passed to EditDistance, causing an OutOfMemoryError.

This now closely matches the similar check in
renderAsFailedStringComparison.