Covering possible int overflow when comparing Strings
Covering a possible int overflow that could occur from int multiplication if MAX_EDIT_DISTANCE_MEMORY was ever increased.

In it's current state, this is unlikely to happen due to the if statement above it filtering out the larger values. However, think it's still useful to "fix" preemptively as is a little bit of a ticking time-bomb if MAX_EDIT_DISTANCE_MEMORY or other depending code is changed in the future.