Fix OOM when comparing large, different Strings (#841)
* Fix OOM when comparing large, very different, Strings

Casting calculation of "edit distance memory" in tryReduceStringSizes to
use longs.

The result of multiplying the length of two Strings can be larger than
the maximum value a Integer can hold before overflowing. In this case,
comparing large strings with big differences could cause an overflow.
When it did overflow, the comparison against MAX_EDIT_DISTANCE_MEMORY
would give an incorrect result, possibly causing large Strings to be
passed to EditDistance, causing an OutOfMemoryError.

This now closely matches the similar check in
renderAsFailedStringComparison.

* Covering possible int overflow when comparing Strings

Covering a possible int overflow that could occur from int multiplication if MAX_EDIT_DISTANCE_MEMORY was ever increased.

In it's current state, this is unlikely to happen due to the if statement above it filtering out the larger values. However, think it's still useful to "fix" preemptively as is a little bit of a ticking time-bomb if MAX_EDIT_DISTANCE_MEMORY or other depending code is changed in the future.

* Improved ExpressionInfoValueRenderer test coverage

Added a test covering a large String comparision, where the size of the difference between the strings only just fits into the MAX_EDIT_DISTANCE_MEMORY limit, leaving no room for a "buffer" around the edge.