Add :spock-specs:mock-integration subproject
The new project contains integration tests for using Spock to mock Java
interfaces/classes with cglib/byte-buddy and objenesis, and without any
of them.

The old testCglib task that reran all tests with byte-buddy
deactivated was removed except for spock-spring which contains a few
tests that only run when cglib is being used.