Fix publishing from Travis CI (#952)
When deploying a release to Nexus, it implicitly creates a staging
repository based on the uploader's IP address. However, when uploading
from a Travis CI build, the external IP address is not stable and thus
Nexus creates multiple staging repositories that can neither be merged
nor released separately in the end.

This commit introduces the `de.marcphilipp.nexus-publish` Gradle plugin
that explicitly creates a staging repository prior to publishing and
uses a URL specific to this repository for uploading. Thus, the auto-
staging feature of Nexus is not used and we're no longer subject to
Travis' issue with external IP addresses.