Fix equality checking of Comparable Mocks (#1322)
`EqualArgumentConstraint` uses `DefaultTypeTransformation.compareEqual`
to check arguments' equality in interactions. For most objects, this is
equivalent to calling `.equals`, so Java Mocks come with an intuitive
default implemention of `.equals`. But for objects that are `Comparable`
(including Mocks of types that implement `Comparable`),
`DefaultTypeTransformation.compareEqual(x, y)` is equivalent to calling
`x.compareTo(y) == 0`. As a result, all Mocks of Comparable types are
treated as equal by `EqualArgumentConstraint`.

This commit fixes that by adding a default implementation of
`.compareTo` to `Comparable` Mocks.