Build MethodSource with the correct spec class name (#1345)
Prior to this commit, all `SpockNode`s that use a `MethodSource`
did not use the actual test class of the discovered method, and 
instead used the declaring class. This was problematic for inherited test
methods, since they appeared to come from the declaring class instead of
the current test class. In addition, the Maven Surefire provider failed
to match such methods when executing tests matching a mask (e.g., via
`-Dtest=*MaskTest`).

Co-authored-by: Miles Thomason <miles.thomason@messagegears.com>
Co-authored-by: Leonard Brünings <lord_damokles@gmx.net>
Co-authored-by: Marc Philipp <marc@gradle.com>