Improve inheritance tests for spec-ignoring annotations
Verify that the feature also works for hierarchies of more than 2
classes, i.e. there are other classes in between annotated spec and
bottom spec.

Also verify that it still works correctly if the parent specs are
non-abstract, i.e. that they are ignored correctly not because
of being abstract but because of the corresponding annotation.

The parent classes also have their own feature methods now, so we can do
a more sophisticated count of executed/skipped features and containers.