Do not evaluate feature-skipping conditions for skipped specs, part 2
Raise the bar in MixedSpecSkippingExtensions, also verifying that
feature-skipping annotation conditions are not evaluated in skipped
parent classes of the bottom spec, up to the actually annotated, skipped
spec.

Before this change, the spec-skipping extensions applied a premature
optimisation I was skeptical of before: to only mark the bottom spec as
skipped for an inherited annotation, instead of marking everything down
to the bottom spec, starting with the annotated spec. This is the proper
way to do it, if we are talking about inheritance.

In order to achieve this more easily, I added new utility methods
'SpecInfo.getSpecsToBottom()' and 'SpecInfo.getSpecsToTop()'. The latter
is unused at the moment, but might come in handy in the future.

Fixes #1458 (really, this time).