Verify that super specs of skipped specs are executed
This has implications for the number of started, skipped, succeeding and
failing tests and containers. We want to make sure the numbers are
reported correctly, looking up the inheritance hierarchy, not just down.
This helps to spot possible regressions where accidentally not just sub
specs, but also or instead super specs would be skipped.