Reset state after intercept of feature.
If @ConfineMetaClassChanges is used on a data-driver feature method, the
ConfineMetaClassChangesInterceptor is shared between iterations of the
method. Clear the originalMetaClasses field of the interceptor after
each invocation to avoid populating the list with duplicate references
to original meta-classes.

The previous implementation was not defective, per se, but this fixes a
potential efficiency issue (determined by how many meta-classes are
stored and how many iterations of the feature method are run).