Do not evaluate feature-skipping conditions for skipped specs (#1459)
The new spec `MixedSpecSkippingExtensions` verifies that this works
correctly for both inherited and non-inherited versions of `@Requires`,
`@IgnoreIf` and - not to forget - `@Ignore`. Because the two feature methods
verify combinations of annotations, they did not fit into any of the
existing specs focused on single annotation-driven extensions. A
separate spec makes more sense. It verifies both the new feature as such
and interoperability between spec-ignoring extension types.

Fixes #1458.

Co-authored-by: Leonard Brünings <lord_damokles@gmx.net>