Update build logic to require JDK 11...
Prior to this commit, the gradle build could be run with JDK 8,
but since a plugin now requires JDK 11, the build needs to be run with JDK 11+.

JDK 8 must still be installed as spock-core is compiled against it via toolchains.
JDK locations should be made known to toolchains via `JDK<version>=<PATH>` environment
variable, e.g., `JDK8=/path/to/jdk8`.