Fix gradle module metadata dependencies
Prior to this commit, the gradle module metadata didn't match the generated pom,
it included more dependencies as Gradle has no concept of provided/optional dependencies.
To fix this we consequently use `compileOnly` for dependencies that are optional.
This will omit these dependencies from both the pom and gmm.

fixes #1469