Make @SpringMock work for beans with multiple instances
fixes #1502

When the context contains multiple beans for given interface
but one of them is marked as @Primary it should be possible
to mock this interface.

Current implementation doesn't check if one of found beans is primary
and fails with an exception.

My change mirrors a similar logic used in spring-boot-starter-test
and resolves this issue.