Remove includedBuild from pluginManagement
Gradle Enterprise complained about different cache settings for the
included builds. This was caused due to the fact that the included build
was referenced in the pluginManagement section, which caused an early
evaluation and different cache settings. As this is only necessary,
for settings plugins, we can just remove it from pluginManagement
to get rid of the warnings.