Fix too many arguments being set as data values in iteration info
The iteration info holds a list of the data values for the current iteration.
The data driver logic had an error where it gave all parameter infos
to the data driver instead of only the ones representing data variables.
Data drivers only deal with data variables, not with additional arguments that
are supplied by extensions. Due to that the data values in the iteration info
had additional MISSING_ARGUMENT entries for the additional arguments which
then also broke things like unroll patterns, conditional annotations accessing data
values and so on