Make @Use on feature and fixture method parallel
For @Use on a specification same-thread policy is necessary as the whole
spec execution is wrapped in a use-closure, so that also helper methods
can use the category methods.

For @Use on features or fixture method this is not necessary, as just the method invocation
is wrapped in the use-closure and this is then always run within one thread.