Log thread dump on unrespected timeout interrupts (#1658)
This PR enhances the `TimeoutInterceptor` to capture thread dumps of the test JVM on unsuccessful interrupt attempts. The thread dump capturing behavior is configurable through the timeout `ConfigurationObject`. Thread dumps are not captured on timeout in favor of communicating the stacktrace of the interrupted thread as part of the `SpockTimeoutError`.