ByteBuddyMockFactory adds @Internal annotation to Groovy MOP methods
The Groovy 3 & 4 runtime expects to have the @Internal annotation on the MOP methods from GroovyObject.
That AbstractCallSite.createGroovyObjectGetPropertySite() processes it as GroovyObject.
So the ByteBuddyMockFactory now adds the @Internal annotation to the intercepted MOP methods.

After that the "Unable to access protected constant when spying instances" problems are gone,
because we take the normal Groovy route.

Also some strange inconsistencies for Groovy  2 <=> 3,4 are now gone,
but a new inconsistency appeared Groovy 4 prefers is over get for boolean.
But this is not a Spock issue, because in Groovy only code the same thing happens.
See tests in JavaMocksForGroovyClasses.

The test "Access protected fields Issue #1452" covers the
"MissingMethodException when accessing fields without explicit @" issue.

This fixes #1145, #1452, #1501, #1608

Co-authored-by: Björn Kautler <Bjoern@Kautler.net>