Interactions in when blocks are not preserved
Interactions in a when: block were not active after the
following then: block, if the then: block contained
an interactions.

Now the SpecRewriter inserts a "freezeScope" after
the moved interactions from then blocks.
This let the MockController skip the scope and
use the outer scope for the interactions from the
when block code.

This fixes #1759