Use multiple locks in ByteBuddyMockFactory to reduce lock contention
Changed locking scheme in ByteBuddyMockFactory from a
single global lock CACHE to cacheLocks with size 16.
The used TypeCachingLock from cacheLocks depends on
the hashcode of TypeCache.SimpleKey, which aggregates the types to mock.

The old global CACHE lock did block Threads regardless,
if they try to mock the same type or not.

This is a similar fix as in Mockito:
https://github.com/mockito/mockito/pull/3095

Co-authored-by: Björn Kautler <Bjoern@Kautler.net>