Move bytecode output to the OUTPUT phase...
in #1765 the behavior of the SourceToAstNodeAndSourceTranspiler was
changed, so that it would output bytecode instructions when the
`CLASS_GENERATION` compile phase was used.
This was problematic as the SourceToAstNodeAndSourceTranspiler is also
used in the Groovy Web Console, and the `CLASS_GENERATION` phase still
had transformations that affected the source and would have not been
accessible anymore. Now we moved the bytecode output trigger to the
`OUTPUT` phase, while internally clamping the actual compiler phase to
the `CLASS_GENERATION` phase.