Defining responses for additionalInterfaces doesn't work
The metaClass used in JavaMocks of Groovy classes did not contain
the methods of additionalInterfaces, so the fix is to change the
MetaClass, if we are a GroovyObject and have additionalInterfaces
to the mocked MetaClass. It is not the best solution, because we now use the
MetaClass of the Mock, which also contains methods and interfaces of Spock
internal, but there is currently no class implementing
mockType + additionalInterfaces. This would be a bigger rework.

Moved the metaClass field down to BaseMockInterceptor
to use it in both cases: GroovyMockInterceptor and JavaMockInterceptor

The second fix is for the JavaMock of a Java class with additionalInterfaces.
Previously the mockType of the Java class was used also for
the methods of the additionalInterfaces, which lead to a NullPointerException.

This fixes #1405