Fix warning about nonspecified root-project name
This warning appears due to the project name of "build-logic" being
inferred from its directory name.
Gradle correctly points out that this can lead to cache issues but
fails to realize that the directory name is part of the git
repository and therefore is unlikely to change.

Specifying the project name explicitely fixes this warning.