ByteBuddy: Do not use Lookup for non-local mocks (#2019)
This PR introduces the concept of local mocks. A mock is considered to
be local if the target class' classloader can also load all additional
interfaces (incl. `ISpockMockObject`).

The lookup classloading strategy may only be used if a mock is local.
Otherwise, we have to defer to other strategies.

Fixes #2017