Configure compiler options via build-logic (#2026)
Moves some of the "uninteresting" options to build-logic to clean up
`<projectname>.gradle` files.

The basic idea of this PR is that much of the tedious configuration
shall happen in the spockframework.base plugin.
All subprojects apply that plugin and profit from the abstraction.

"Interesting stuff" like declaration of dependencies shall still happen
in `<projectname>.gradle` files.
I'd like to expand this idea , but I want to keep the PRs small and
start with the most simple stuff first.